+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 **English 🇺🇸️** / [Deutsch 🇩🇪️](CHANGELOG.de.md)
 
+### Version 1.2.0
+Added:
+- Cross-validation `-loocv`
+- Option `-altitude`
+- Data sheets from Cleveland 1927–1929
+- Adjustments to process the data sheets from Cleveland 1927–1929
+
+Changed:
+- Removed directory Shankland1955
+
+
 ### Version 1.1.0
 Added:
 - Filter `-abs_drift`
@@ -15,6 +26,7 @@ Changed:
 Fixed:
 Berichtigt:
 - In `data_sheets.ods` and thus also in `data_sheets.pdf` some wrong characters were removed in the columns _sign_ and _sign correct_.
+
 
 ### Version 1.0.0
 Added: